64bit bugfixes

+++ include/ucui.h
@@ -133,7 +133,7 @@ void printwass(unsigned int startpos, unsigned int endpos, uint64_t pc);
 struct disassembly * disass(uint8_t *code, unsigned int len, uint64_t baseaddress, cs_arch arch, cs_mode mode);
 struct readfile * readfile(char *filename);
 void verify_visible_ip(uint32_t pc);
-bool ip_aligned_to_disassembly(uint32_t pc);
+bool ip_aligned_to_disassembly(uint64_t pc);
 bool should_break(uint64_t pc);
 void redisassemble_code(uc_engine *uc, uint64_t ip, size_t len);
